+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Stream recorded shader frames to GIF or APNG with the optional Flipbook gem.
+
 ## [0.1.0] - 2026-09-23
 
 - Initial release.